requirements.txt
Django==6.0.4
python-decouple==3.8
django-crispy-forms==2.3
crispy-tailwind==0.5.0
django-cors-headers==4.4.0
djangorestframework==3.15.0
django-filter==24.1
django-extensions==3.2.3
whitenoise==6.8.2
python-dotenv==1.0.1
django-model-utils==4.5.1
django-taggit==6.1.0
dj-database-url==2.1.0

